satisfy the type checker

+++ posthog/test/test_migration_0227.py
@@ -71,8 +71,9 @@ class CreatingDashboardTilesTestCase(TestMigrations):
         self.assertEqual(DashboardTile.objects.filter(dashboard__name="d3").count(), 0)
 
     def tearDown(self):
-        Team = self.apps.get_model("posthog", "Team")
-        Dashboard = self.apps.get_model("posthog", "Dashboard")
+        if self.apps is not None:
+            Team = self.apps.get_model("posthog", "Team")
+            Dashboard = self.apps.get_model("posthog", "Dashboard")
 
-        Dashboard.objects.all().delete()
-        Team.objects.all().delete()
+            Dashboard.objects.all().delete()
+            Team.objects.all().delete()
